refactor: implement policy registry and detector orchestration
